code/proto/normalize_arxiv_dump_requirements.txt
chardet>=3.0.4
python-magic>=0.4.15
